still need to fix refraction and reflection

+++ Source/Engine/Renderer/Texture.cpp
@@ -38,8 +38,8 @@ namespace neu {
         GLenum internalFormat = (channels == 4) ? GL_RGBA8 : GL_RGB8;
         GLenum format = (channels == 4) ? GL_RGBA : GL_RGB;
 
-		m_size.x = (float)surface->w;
-		m_size.y = (float)surface->h;
+		m_size.x = surface->w;
+		m_size.y = surface->h;
 
         glGenTextures(1, &m_texture);
         glBindTexture(m_target, m_texture);
